+++ lib/utils/extensions/navigation_extension.dart
@@ -2,10 +2,23 @@ import 'package:dartx/dartx.dart';
 import 'package:flutter/material.dart';
 
 extension NavigationExtension on BuildContext {
-  Future<void> navigateTo(Widget destination,
-      {Duration delay = const Duration(), bool replace = false}) async {
+  Future<void> navigateTo(
+    Widget destination, {
+    Duration delay = const Duration(),
+    bool replace = false,
+  }) async {
     await delay.delay;
-    final route = PageRouteBuilder(
+    final route = _createRoute(destination);
+
+    if (replace) {
+      Navigator.of(this).pushReplacement(route);
+    } else {
+      Navigator.of(this).push(route);
+    }
+  }
+
+  PageRouteBuilder _createRoute(Widget destination) {
+    return PageRouteBuilder(
       pageBuilder: (context, animation, secondaryAnimation) => destination,
       transitionsBuilder: (context, animation, secondaryAnimation, child) {
         const begin = Offset(1.0, 0.0);
@@ -15,17 +28,8 @@ extension NavigationExtension on BuildContext {
         final tween =
             Tween(begin: begin, end: end).chain(CurveTween(curve: curve));
 
-        return SlideTransition(
-          position: animation.drive(tween),
-          child: child,
-        );
+        return SlideTransition(position: animation.drive(tween), child: child);
       },
     );
-
-    if (replace) {
-      Navigator.of(this).pushReplacement(route);
-    } else {
-      Navigator.of(this).push(route);
-    }
   }
 }
